:recycle: (Auth) Extract logInButton from header
